requirements.txt
azure-functions
pandas
azure-storage-blob
scikit-learn